coeff

+++ wl-bot/src/utils/comparator/desc/RegionWeightPickDescComparator.java
@@ -8,7 +8,7 @@ public class RegionWeightPickDescComparator implements Comparator<Region> {
 
 	@Override
 	public int compare(Region region1, Region region2) {
-		return Integer.valueOf(region2.getWeightPick()).compareTo(Integer.valueOf(region1.getWeightPick()));
+		return Double.valueOf(region2.getWeightPick()).compareTo(Double.valueOf(region1.getWeightPick()));
 	}
 
 }
